refactor: move safeJsonParse to json-utils and update imports
