+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/history/event-log/event-log-table.component.ts
@@ -2,7 +2,7 @@ import { formatDate } from '@angular/common';
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { CaseViewEvent } from '../../../../domain';
 import { SessionStorageService } from '../../../../services';
-import { safeJsonParse } from '../../../../utils';
+import { safeJsonParse } from '../../../../json-utils';
 
 @Component({
   selector: 'ccd-event-log-table',
