+++ projects/ccd-case-ui-toolkit/src/lib/shared/index.ts
@@ -7,3 +7,4 @@ export * from './services';
 export * from './test/test-route-snapshot-builder';
 export * from './services/fields';
 export * from './commons/case-edit-data';
+export * from './json-utils';
