+++ projects/ccd-case-ui-toolkit/src/lib/shared/services/session/session-storage.service.ts
@@ -6,7 +6,7 @@ export class SessionStorageService {
   /**
    * Get an item from the session storage.
    */
-  public getItem(key: string): string {
+  public getItem(key: string): string | null {
     return sessionStorage.getItem(key);
   }
