+++ projects/ccd-case-ui-toolkit/src/lib/shared/utils.spec.ts
@@ -1,4 +1,4 @@
-import { safeJsonParse } from './utils';
+import { safeJsonParse } from './json-utils';
 
 describe('safeJsonParse', () => {
   it('returns fallback when value is null', () => {
